Support configuring AUX 5 to 8.
The MSP is changed in a way that might provide some backwards
compatibility.  The first 4 channels are sent/read as before followed by
the next 4 channels.

If I client ignores extra data received it should be backwards
compatible.

Clients can looks for the new capability bit which indicates the MSP
protocol supports AUX 1-8.

+++ src/main/config/config_profile.h
@@ -42,7 +42,7 @@ typedef struct profile_s {
 
     uint8_t acc_unarmedcal;                 // turn automatic acc compensation on/off
 
-    uint16_t activate[CHECKBOX_ITEM_COUNT]; // activate switches
+    uint32_t activate[CHECKBOX_ITEM_COUNT]; // activate switches, bitmask, 3 bits per channel, lower 16 bits aux1-4, upper 16 bits aux 5-8
 
     // Radio/ESC-related configuration
     uint8_t deadband;                       // introduce a deadband around the stick center for pitch and roll axis. Must be greater than zero.
